# openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/products:
    post:
      operationId: ProductsController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Create a product
      tags: &a15
        - products
    get:
      operationId: ProductsController_findAll_v1
      parameters:
        - name: take
          required: false
          in: query
          schema:
            maximum: 100
            default: 100
            type: number
        - name: skip
          required: false
          in: query
          schema:
            default: 0
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: List the consumer's products
      tags: *a15
  /v1/products/{id}:
    get:
      operationId: ProductsController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Get a product by id
      tags: *a15
    patch:
      operationId: ProductsController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Update a product
      tags: *a15
    delete:
      operationId: ProductsController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDeletedEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Delete a product
      tags: *a15
...
info:
  title: Cosmos Pay — Payments API
  description: "Payments microservice (Stellar payment intents). All endpoints
    require traffic to arrive through the APISIX gateway: a valid
    `X-Gateway-Secret` header plus an authenticated consumer
    (`X-Consumer-Username`). Paths already include the version (`/v1/...`)."
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    gateway-secret:
      type: apiKey
      in: header
      name: X-Gateway-Secret
    consumer:
      type: apiKey
      in: header
      name: X-Consumer-Username
  schemas:
...
    CreateProductDto:
      type: object
      properties:
        name:
          type: string
          example: Pro plan — monthly
        description:
          type: string
          example: Full access, billed every month.
        amount:
          type: string
          example: "49.00"
          description: Decimal price (≤ 7 decimals). Omit for a customer-set amount.
        assetCode:
          type: string
          example: USDC
          description: Asset code the price is in. Omit for native lumens (XLM).
        kind:
          type: string
          enum: &a24
            - recurring
            - one_time
            - link
          default: one_time
        active:
          type: boolean
          default: true
        reference:
          type: string
          example: sku_pro_monthly
      required:
        - name
    ProductEntity:
      type: object
      properties:
        id:
          type: string
          example: clx9z8a1b0000abcd1234efgh
        name:
          type: string
          example: Pro plan — monthly
        description:
          type: object
          nullable: true
          example: Full access, billed every month.
        amount:
          type: object
          nullable: true
          example: "49.00"
        asset:
          type: string
          example: native
          description: Asset code, or "native" for XLM.
        kind:
          type: string
          example: one_time
          enum:
            - recurring
            - one_time
            - link
        active:
          type: boolean
          example: true
        reference:
          type: object
          nullable: true
          example: sku_pro_monthly
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - name
        - description
        - amount
        - asset
        - kind
        - active
        - reference
        - createdAt
        - updatedAt
    ProductListEntity:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProductEntity"
        total:
          type: number
          example: 137
          description: Total products matching the query, not the page size.
        take:
          type: number
          example: 100
        skip:
          type: number
          example: 0
      required:
        - data
        - total
        - take
        - skip
    UpdateProductDto:
      type: object
      properties:
        name:
          type: string
          example: Pro plan — monthly
        description:
          type: string
          example: Full access, billed every month.
        amount:
          type: string
          example: "49.00"
          description: Decimal price (≤ 7 decimals). Omit for a customer-set amount.
        assetCode:
          type: string
          example: USDC
          description: Asset code the price is in. Omit for native lumens (XLM).
        kind:
          type: string
          enum: *a24
          default: one_time
        active:
          type: boolean
          default: true
        reference:
          type: string
          example: sku_pro_monthly
    ProductDeletedEntity:
      type: object
      properties:
        id:
          type: string
          example: clx9z8a1b0000abcd1234efgh
        deleted:
          type: boolean
          example: true
      required:
        - id
        - deleted
...
security:
  - gateway-secret: []
  - consumer: []